+++ src/core/util/parallel.h
@@ -26,7 +26,8 @@ inline void free_rand() { }
 inline int calc_threads(int computations) { return 0; }
 inline int calc_blocks(int computations, int threads=0) { return 0; }
 inline void device_check_memory() { }
-inline void* cuda_allocate_device(int device_id, int count, int size, void* source_data) {
+inline void* cuda_allocate_device(int device_id, unsigned long count,
+        int size, void* source_data) {
     ErrorManager::get_instance()->log_error(
         "Attempted to allocate device memory in non-parallel build.");
     return nullptr;
@@ -62,7 +63,8 @@ void gpuAssert(const char* file, int line, const char* msg);
 
 /** Checks cuda memory usage and availability */
 void device_check_memory();
-void* cuda_allocate_device(int device_id, int count, int size, void* source_data);
+void* cuda_allocate_device(int device_id, unsigned long count,
+        int size, void* source_data);
 
 // Random state data
 extern DEVICE curandState_t* cuda_rand_states;
